Extracted Map class out of the MapManager class.
